Move targets for multiprocessing in snowflake
